Rename chartUtils to chartDatumUtils
