+++ src/components/charts/common/chartDatumUtils.test.ts
@@ -1,7 +1,7 @@
 jest.mock('i18next');
 import i18next from 'i18next';
 
-import { getTooltipContent } from './chartUtils';
+import { getTooltipContent } from './chartDatumUtils';
 
 describe('getTooltipContent', () => {
   test('format hrs and gb', () => {
