+++ src/components/charts/common/chartLegendItem.test.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { mockDate } from 'testUtils';
 
 import { ChartLegendItem, ChartLegendItemProps } from './chartLegendItem';
-import * as utils from './chartUtils';
+import * as utils from './chartDatumUtils';
 
 mockDate();
