+++ src/components/charts/common/index.ts
@@ -1,4 +1,4 @@
 export { ChartLegend, ChartLegendProps } from './chartLegend';
 export { ChartLegendItem, ChartLegendItemProps } from './chartLegendItem';
 export { ChartTitle, ChartTitleProps } from './chartTitle';
-export { ChartDatum } from './chartUtils';
+export { ChartDatum } from './chartDatumUtils';
