+++ src/components/charts/usageChart/usageChart.tsx
@@ -17,7 +17,7 @@ import {
   getMaxMinValues,
   getTooltipContent,
   getUsageRangeString,
-} from 'components/charts/common/chartUtils';
+} from 'components/charts/common/chartDatumUtils';
 import getDate from 'date-fns/get_date';
 import i18next from 'i18next';
 import React from 'react';
