+++ src/store/dashboard/awsCloudDashboard/awsCloudDashboardWidgets.ts
@@ -4,7 +4,7 @@ import {
   ComputedForecastItemType,
   ComputedReportItemType,
   ComputedReportItemValueType,
-} from 'components/charts/common/chartUtils';
+} from 'components/charts/common/chartDatumUtils';
 import { DashboardChartType } from 'store/dashboard/common/dashboardCommon';
 
 import { ForecastPathsType, ForecastType } from '../../../api/forecasts/forecast';
