+++ src/store/dashboard/azureCloudDashboard/azureCloudDashboardWidgets.ts
@@ -5,7 +5,7 @@ import {
   ComputedForecastItemType,
   ComputedReportItemType,
   ComputedReportItemValueType,
-} from 'components/charts/common/chartUtils';
+} from 'components/charts/common/chartDatumUtils';
 import { DashboardChartType } from 'store/dashboard/common/dashboardCommon';
 
 import { AzureCloudDashboardTab, AzureCloudDashboardWidget } from './azureCloudDashboardCommon';
